server-ai/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.3
langchain==0.2.1
openai==1.30.5
python-dotenv==1.0.1
